requirements.txt
# Plant Health Evaluation Framework
# Install with: pip install -r requirements.txt

# Testing
pytest>=7.0.0

# LLM Integration (Gemini)
google-genai>=1.57.0

# Chat UI
chainlit>=1.0.0

# Environment management
python-dotenv>=1.0.0

# PII Detection
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Observability
arize-phoenix>=4.0.0
openinference-instrumentation-google-genai>=0.1.0